requirements.txt.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.24.0
Pillow
matplotlib
tqdm
